backend/analytics-service/requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
sqlalchemy==2.0.46
psycopg2-binary==2.9.11
python-jose[cryptography]==3.5.0
python-dotenv==1.2.1
pydantic==2.12.5
pydantic-settings==2.12.0
